feat: added user authentication with firebase
Implemented user authentication with a session provider and firebase sdk. Includes a major reorganization of the routing file paths by making signin
a standalone file and using a protected path in the form of (app)/*. Upon successful signin, users are redirected to the protected route.

Exposes hardcoded API keys/values that will need to be obscured from the user in production.
