+++ package.json
@@ -33,6 +33,7 @@
     "expo-symbols": "~0.2.0",
     "expo-system-ui": "~4.0.2",
     "expo-web-browser": "~14.0.1",
+    "firebase": "^11.0.2",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-native": "0.76.1",
